+++ internal/cache.go
@@ -264,7 +264,7 @@ func (cache *Cache) FetchTwts(conf *Config, archive Archiver, feeds types.Feeds,
 				twter := types.Twter{Nick: feed.Nick}
 				if strings.HasPrefix(feed.URL, conf.BaseURL) {
 					twter.URL = URLForUser(conf, feed.Nick)
-					twter.Avatar = URLForAvatar(conf, feed.Nick)
+					twter.Avatar = URLForAvatar(conf.BaseURL, feed.Nick)
 				} else {
 					twter.URL = feed.URL
 					avatar := GetExternalAvatar(conf, feed.Nick, feed.URL)
