+++ internal/handlers.go
@@ -18,7 +18,6 @@ import (
 	"time"
 
 	"github.com/chai2010/webp"
-	securejoin "github.com/cyphar/filepath-securejoin"
 	"github.com/dgrijalva/jwt-go"
 	"github.com/gomarkdown/markdown"
 	"github.com/gomarkdown/markdown/html"
@@ -430,27 +429,6 @@ func (s *Server) ArchiveFeedHandler() httprouter.Handle {
 	}
 }
 
-// OldTwtxtHandler ...
-// Redirect old URIs (twtxt <= v0.0.8) of the form /u/<nick> -> /user/<nick>/twtxt.txt
-// TODO: Remove this after v1
-func (s *Server) OldTwtxtHandler() httprouter.Handle {
-	return func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-		nick := NormalizeUsername(p.ByName("nick"))
-		if nick == "" {
-			http.Error(w, "Bad Request", http.StatusBadRequest)
-			return
-		}
-
-		newURI := fmt.Sprintf(
-			"%s/user/%s/twtxt.txt",
-			strings.TrimSuffix(s.config.BaseURL, "/"),
-			nick,
-		)
-
-		http.Redirect(w, r, newURI, http.StatusMovedPermanently)
-	}
-}
-
 // OldAvatarHandler ...
 // Redirect old URIs (twtxt <= v0.1.0) of the form /user/<nick>/avatar.png -> /user/<nick>/avatar
 // TODO: Remove this after v1
@@ -588,105 +566,6 @@ func (s *Server) AvatarHandler() httprouter.Handle {
 			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
 			return
 		}
-	}
-}
-
-// TwtxtHandler ...
-func (s *Server) TwtxtHandler() httprouter.Handle {
-	return func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-		nick := NormalizeUsername(p.ByName("nick"))
-		if nick == "" {
-			http.Error(w, "Bad Request", http.StatusBadRequest)
-			return
-		}
-
-		fn, err := securejoin.SecureJoin(filepath.Join(s.config.Data, "feeds"), nick)
-		if err != nil {
-			http.Error(w, "Bad Request", http.StatusBadRequest)
-			return
-		}
-
-		fileInfo, err := os.Stat(fn)
-		if err != nil {
-			if os.IsNotExist(err) {
-				http.Error(w, "Feed Not Found", http.StatusNotFound)
-				return
-			}
-
-			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-			return
-		}
-
-		w.Header().Set("Content-Type", "text/plain; charset=utf-8")
-		w.Header().Set("Content-Length", fmt.Sprintf("%d", fileInfo.Size()))
-		w.Header().Set("Link", fmt.Sprintf(`<%s/user/%s/webmention>; rel="webmention"`, s.config.BaseURL, nick))
-		w.Header().Set("Last-Modified", fileInfo.ModTime().UTC().Format(http.TimeFormat))
-
-		followerClient, err := DetectFollowerFromUserAgent(r.UserAgent())
-		if err != nil {
-			log.WithError(err).Warnf("unable to detect twtxt client from %s", FormatRequest(r))
-		} else {
-			var (
-				user       *User
-				feed       *Feed
-				err        error
-				followedBy bool
-			)
-
-			if user, err = s.db.GetUser(nick); err == nil {
-				followedBy = user.FollowedBy(followerClient.URL)
-			} else if feed, err = s.db.GetFeed(nick); err == nil {
-				followedBy = feed.FollowedBy(followerClient.URL)
-			} else {
-				log.WithError(err).Warnf("unable to load user or feed object for %s", nick)
-			}
-
-			if (user != nil) || (feed != nil) {
-				if (s.config.Debug || followerClient.IsPublicURL()) && !followedBy {
-					if _, err := AppendSpecial(
-						s.config, s.db,
-						twtxtBot,
-						fmt.Sprintf(
-							"FOLLOW: @<%s %s> from @<%s %s> using %s",
-							nick, URLForUser(s.config, nick),
-							followerClient.Nick, followerClient.URL,
-							followerClient.Client,
-						),
-					); err != nil {
-						log.WithError(err).Warnf("error appending special FOLLOW post")
-					}
-
-					if user != nil {
-						user.AddFollower(followerClient.Nick, followerClient.URL)
-						if err := s.db.SetUser(nick, user); err != nil {
-							log.WithError(err).Warnf("error updating user object for %s", nick)
-						}
-					} else if feed != nil {
-						feed.AddFollower(followerClient.Nick, followerClient.URL)
-						if err := s.db.SetFeed(nick, feed); err != nil {
-							log.WithError(err).Warnf("error updating feed object for %s", nick)
-						}
-					} else {
-						panic("should not be reached")
-						// Should not be reached
-					}
-				}
-			}
-		}
-
-		f, err := os.Open(fn)
-		if err != nil {
-			log.WithError(err).Error("error opening feed")
-			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-			return
-		}
-		defer f.Close()
-
-		if r.Method == http.MethodHead {
-			return
-		}
-
-		http.ServeContent(w, r, filepath.Base(fn), fileInfo.ModTime(), f)
 	}
 }
 
@@ -925,7 +804,7 @@ func (s *Server) PermalinkHandler() httprouter.Handle {
 
 		if isLocal(twt.Twter().URL) {
 			who = fmt.Sprintf("%s@%s", twt.Twter().Nick, s.config.baseURL.Hostname())
-			image = URLForAvatar(s.config, twt.Twter().Nick)
+			image = URLForAvatar(s.config.BaseURL, twt.Twter().Nick)
 		} else {
 			who = fmt.Sprintf("@<%s %s>", twt.Twter().Nick, twt.Twter().URL)
 			image = URLForExternalAvatar(s.config, twt.Twter().URL)
