+++ types/profile.go
@@ -4,11 +4,12 @@ package types
 type Profile struct {
 	Type string
 
-	Username string
-	Tagline  string
-	URL      string
-	TwtURL   string
-	BlogsURL string
+	Username  string
+	Tagline   string
+	URL       string
+	TwtURL    string
+	BlogsURL  string
+	AvatarURL string
 
 	// `true` if the User viewing the Profile has muted this user/feed
 	Muted bool
